ACAN_ESP32: Port TWAI_ERRATA_FIX_BUS_OFF_REC controller errata
When the bus-off condition is reached, the REC should be reset to 0
and frozen (via LOM) by the driver's ISR.

However on the ESP32, there is an edge case where the REC will increase
before the driver's ISR can respond in time (e.g., due to the rapid
occurrence of bus errors), thus causing the REC to be non-zero after
bus-off.

A non-zero REC can prevent bus-off recovery as the bus-off recovery
condition is that both TEC and REC become 0.

Add a workaround in the driver to forcibly reset REC to zero on
reaching bus-off.

The errata is ported from the espressif ESP-IDF CAN driver.
https://github.com/espressif/esp-idf/commit/2f5806092135e3d991057bc06225bdcf536e93a5