ACAN_ESP32: Port TX interrupt lost errata
On the ESP32, when a transmit interrupt occurs, and interrupt register
is read on the same APB clock cycle, the transmit interrupt could be
lost.

Add a workaround that checks the transmit buffer status bit to recover
any lost transmit interrupt.

The errata is ported from the espressif ESP-IDF CAN driver.
https://github.com/espressif/esp-idf/commit/2f5806092135e3d991057bc06225bdcf536e93a5